+++ src/BlockHorizons/BlockSniper/changelog/ChangeLog.php
@@ -1,0 +1,76 @@
+<?php
+
+declare(strict_types=1);
+
+namespace BlockHorizons\BlockSniper\changelog;
+
+use BlockHorizons\BlockSniper\ui\forms\ModalForm;
+use pocketmine\utils\TextFormat;
+
+class ChangeLog{
+
+	/** @var ChangeLog[] */
+	public static $changeLogs = [];
+
+	/** @var string */
+	private $version;
+	/** @var string  */
+	private $date;
+
+	/** @var string[] */
+	private $added, $changed, $removed, $fixed;
+
+	public function __construct(string $version, string $date, array $added, array $changed, array $removed, array $fixed){
+		$this->date = $date;
+		$this->version = $version;
+		$this->added = $this->filter($added);
+		$this->changed = $this->filter($changed);
+		$this->removed = $this->filter($removed);
+		$this->fixed = $this->filter($fixed);
+	}
+
+	/**
+	 * @return ModalForm
+	 */
+	public function toForm() : ModalForm {
+		$text = "";
+		foreach(["added" => $this->added, "changed" => $this->changed, "removed" => $this->removed, "fixed" => $this->fixed] as $key => $changes){
+			switch($key){
+				case "added": $text .= TextFormat::GREEN; break;
+				case "changed": $text .= TextFormat::GOLD; break;
+				case "removed": $text .= TextFormat::RED; break;
+				case "fixed": $text .= TextFormat::AQUA; break;
+			}
+			$text .= TextFormat::BOLD . ucfirst($key) . TextFormat::RESET . "\n";
+
+			foreach($changes as $change){
+				if($change[0] === "-"){
+					// Wordwrap it with indentation so that everything is indented properly.
+					$text .= wordwrap($change, 46, "\n  ");
+				} else {
+					// Wordwrap it with even more indentation as this change is indented 4 spaces (or a tab) further.
+					$text .= wordwrap(str_replace("-", "o", $change), 42, "\n       ");
+				}
+				$text .= "\n";
+			}
+			$text .= "\n";
+		}
+		$form = new ModalForm("BlockSniper $this->version Changelog ", $text);
+		return $form;
+	}
+
+	/**
+	 * @param string[] $changes
+	 *
+	 * @return string[]
+	 */
+	private function filter(array $changes) : array {
+		$new = [];
+		foreach($changes as $key => $value) {
+			if(trim($value) !== ""){
+				$new[$key] = $value;
+			}
+		}
+		return $new;
+	}
+}